Move some common code out to its own library folder
