+++ platformio.ini
@@ -121,6 +121,8 @@ build_flags =
 	-O0
 build_unflags =
 	-Os
+lib_extra_dirs =
+	src_common
 
 [env:test]
 platform = native@1.2.1
@@ -128,6 +130,8 @@ build_flags =	-std=gnu++11  -include test/mock.h
 lib_deps =
 	ArduinoFake
 test_build_src = true
+lib_extra_dirs =
+	src_common
 
 [env:webserver]
 platform = https://github.com/platformio/platform-espressif32.git
@@ -140,6 +144,7 @@ custom_src_dir = webserver
 build_src_filter = +<*> -<web-interface>
 
 build_flags =
+	-std=gnu++17
 	-DBOARD_HAS_PSRAM
 	-mfix-esp32-psram-cache-issue
 
@@ -147,7 +152,11 @@ lib_deps =
 	https://github.com/me-no-dev/AsyncTCP.git
 	https://github.com/me-no-dev/ESPAsyncWebServer.git
 	ArduinoJSON@^6.19.4
+	Wire
 
 extra_scripts =
 	pre:scripts/custom-src-dir.py
 	webserver/replace_fs.py
+
+lib_extra_dirs =
+	src_common
