+++ src/functional/just_do_coffee.h
@@ -1,11 +1,11 @@
 #ifndef JUST_DO_COFFEE_H
 #define JUST_DO_COFFEE_H
 
-#include "../utils.h"
+#include "utils.h"
 #include "../peripherals/peripherals.h"
 #include "../peripherals/pump.h"
 #include "../eeprom_data/eeprom_data.h"
-#include "../sensors_state.h"
+#include "sensors_state.h"
 #include <Arduino.h>
