+++ src/gaggiuino.h
@@ -14,7 +14,7 @@
 #include "functional/descale.h"
 #include "functional/just_do_coffee.h"
 #include "functional/predictive_weight.h"
-#include "functional/profiling_phases.h"
+#include "profiling_phases.h"
 
 #include <Arduino.h>
 #include <SimpleKalmanFilter.h>
