+++ src_common/common_code/include/sensors_state.h
