+++ src_common/common_code/library.json
@@ -1,0 +1,6 @@
+{
+  "name": "common_code",
+  "version": "0.0.1",
+  "keywords": "",
+  "description": "Contains common structures and comms code for UART comms of Gaggiuino data"
+  }
