+++ src_common/common_code/src/utils.cpp
